Experimental proof-of-concept of contextual autoescaping in PUG
Make it easier to write XSS-free PUG frontends by removing a couple
of major XSS vectors

1.  URL attribute values that are not constant expressions are now
    filtered so that `javascript:` URLs will not cause execution.
    This is the single largest XSS vector after injecting `<script>`
2.  `&attributes(...)` are now checked so that an attacker who
    controls a key and value cannot inject JavaScript via
    event handlers attributes.
3.  Text interpolation into `<script>` elements is JSON escaped.

For example, given the template

```pug
//-
  Use an opaque predicate to trick Constantinople into treating this as
  requiring evaluation at render time
- var js = Math.random() <= 2 ? "javascript:alert(1)//" : ""
- var attributes = {}
- attributes.href = js + '/'

p
  a(href='javascript:alert(1)') Link with constant href
p
  a&attributes({ href: 'javascript:alert(1)' }) Constant href via ampersand
  | TODO: This is not ideal.
p
  <a href="javascript:alert(1)">
    | Static link not in pug syntax
  </a>
p
  <a href="#{js}">
    | Dynamic link not in pug syntax
  </a>
ul
  li
    a(href=js + '/') Concatenation
  li
    a(href=`${js}/`) Template string
  li
    a(href!=js + '/') Unescaped concatenation
  li
    a&attributes({href: js + '/'}) Ampersand with map constructor
  li
    a&attributes(attributes) Ampersand with expression
```

without this change results in

```html
<p><a href=\"javascript:alert(1)\">Link with constant href</a></p>
<p><a href=\"javascript:alert(1)\">Constant href via ampersand</a>TODO: This is not ideal.</p>
<p>
  <a href=\"javascript:alert(1)\">Static link not in pug syntax</a></p>
<p><a href=\"javascript:alert(1)//\">Dynamic link not in pug syntax</a></p>
<ul>
  <li><a href=\"javascript:alert(1)///\">Concatenation</a></li>
  <li><a href=\"javascript:alert(1)///\">Template string</a></li>
  <li><a href=\"javascript:alert(1)///\">Unescaped concatenation</a></li>
  <li><a href=\"javascript:alert(1)///\">Ampersand with map constructor</a></li>
  <li><a href=\"javascript:alert(1)///\">Ampersand with expression</a></li>
</ul>
```

but with these changes produces

```html
<p><a href="javascript:alert(1)">Link with constant href</a></p>
<p><a href="#zSafehtmlz">Constant href via ampersand</a>TODO: This is not ideal.</p>
<p>
  <a href="javascript:alert(1)">Static link not in pug syntax</a></p>
<p><a href="#zSafehtmlz">Dynamic link not in pug syntax</a></p>
<ul>
  <li><a href="#zSafehtmlz">Concatenation</a></li>
  <li><a href="#zSafehtmlz">Template string</a></li>
  <li><a href="#zSafehtmlz">Unescaped concatenation</a></li>
  <li><a href="#zSafehtmlz">Ampersand with map constructor</a></li>
  <li><a href="#zSafehtmlz">Ampersand with expression</a></li>
</ul>
```

This adds a `packages/pug-autoesc` which implements a contextual
autoescaper similar to:

*  [Angular SCE](https://docs.angularjs.org/api/ng/service/$sce)
*  [Closure Templates](https://developers.google.com/closure/templates/docs/security)
*  [golang `html/template`](https://golang.org/pkg/html/template/)
*  [polymer-resin](https://github.com/Polymer/polymer-resin)

It modifies the code generating code in `pug-code-gen` and `pug-attrs`
to consolidate code that appends output to `pug_html` so that, when
`options.bufferHooksExperimental` is present, the generated code
instead can write to a buffer object.

`packages/pug-autoesc` defines such a buffer class.

It modifies `packages/pug-runtime` to make the buffer class and
a few supporting functions available at runtime.

Similar systems allow type-safe exceptions to the rule via typed
strings like Angular's [SafeValue][] type as detailed in ["Securing
the Tangled Web"][].

Should the runtime check if `require('@angular/platform-browser')`
succeeds at runtime and treat such values as privileged when
such values are written at runtime.

["Securing the Tangled Web"]: https://research.google.com/pubs/pub42934.html
SafeValue: https://angular.io/api/platform-browser/SafeValue

----

Once we have context about which scripts and styles are under
template author control, we can also make it easier to deploy
the templates under a [Content-Security-Policy][].
Would PUG users benefit from [auto-noncing][]?

[Content-Security-Policy]: https://developers.google.com/web/fundamentals/security/csp/
[auto-noncing]: https://developers.google.com/closure/templates/docs/security#content_security_policy

Pug has a [plugin mechanism][] that can operate on the source files,
token stream, ASTs, or generated JS.

This proof-of-concept is not written as a plugin.  Instead it
uses hooks so that generated code of the form
(`pug_html = pug_html + ...`) can be replaced with writes
that accumulate content onto a buffer along with information
about whether that content comes from the template author or
not.

It might be possible to implement equivalent functionality as an
AST plugin that adds calls to filter functions to the AST.
Doing this conservatively might require complicated static analysis
that is aware of template inheritance and which takes into account
the contexts in which mixins are called.

If this approach is taken, the AST pass first need to translate
constructs like

```pug
<a href="#{...}">...
```

to

```pug
a(href=`$(...)`)
```

or just reject them, and behave similarly on unescaped buffered code
interpolations like

```pug
p
  != thereBeTagsOfUnkownProvenanceHere
```

Doing things this way would make the functioning of the plugin
dependent on any later AST processing plugins, and it would have to
panic if there are new kinds of AST nodes added to be conservative.

[plugin mechanism]: https://github.com/pugjs/pug/blob/master/packages/pug/lib/index.js#L43